[FE] refactor: move useRef
Move useRef to presentational layer
